update 衝突処理をCubeCollisionに変更。Cube側にアタッチしてその処理だけで完結するように変更
update
不要な行を削除
